fix bug by updating the labels/websites passed in init.js. Move getPermMapping to filterLabels.js
